+++ index.js
@@ -1,5 +1,0 @@
-export {
-    EventMap,
-    event,
-    eventsListenAt
-} from "@harbor/eventmap";