+++ packages/Middleware/index.ts
@@ -1,8 +1,0 @@
-export {
-  Logger,
-  loggerConfig
-} from "./Logger";
-
-export {
-  Middleware
-} from "./src/Middleware";