+++ CHANGELOG.md
@@ -5,6 +5,94 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+## Updated
+- StateChange package 1.0.0
+- testUtils package 1.0.0
+
+## [0.21.0] 2021-09-30
+### Updated
+- Router package 0.3.0
+
+## [0.20.0] 2021-09-29
+### Updated
+- Router package 0.2.0
+- DataElement 0.5.2
+- StateChange 0.7.2
+
+
+## [0.19.0] 2021-09-22
+### Added
+- Router package 0.1.1
+### Updated
+ - Event Map 0.7.1
+ - DataElement 0.5.1
+ - StateChange 0.7.1
+ - linkProp 0.3.1
+ - testUtils 0.2.1
+ - Middleware 0.2.9
+ - functional 0.2.7
+
+
+## [0.18.0] - 2021-09-10
+### Updated
+- badge inclusion and test coverage scripts
+- All packages patch version for documentation update
+
+## [0.17.0] - 2021-09-09
+### Updated
+- DataElements package to 0.4.0
+
+## [0.16.0] - 2021-09-08
+### Updated
+- DataElements package to 0.3.0
+
+## [0.15.0] - 2021-09-07
+### Updated
+- DataElements package to 0.2.0
+
+## [0.14.0] - 2021-09-06
+### Added
+- DataElements package 0.1.0
+
+## [0.13.0] - 2021-09-04
+### Added
+- testUtils package 0.1.0
+### Updated
+- EventMap package to 0.6.0
+
+
+## [0.12.0] - 2021-09-04
+### Updated
+- StateChange package to 0.6.0
+
+
+## [0.11.0] - 2021-09-03 
+### Updated
+- StateChange package to 0.5.0
+- EventMap package to 0.5.0
+
+## [0.10.0] - 2021-08-12
+### Updated
+- EventMap package to 0.4.0
+- StateChange package to 0.4.0
+
+## [0.9.0] - 2021-08-07
+### Fixed
+- EventMap bug
+### Changed
+- StateChange middleware method names
+- Package contents
+
+## [0.8.0] - 2021-08-05
+### Updated
+- Changes to packages after publishing
+
+## [0.7.0] - 2021-08-05
+### Changed
+- Updated dependencies
+- Published packages
+
 ## [0.6.0] - 2021-08-04
 ### Updated
 - Updated StateChange package to 0.3.0